La variable n debía ser float.

+++ 2.C++/TPN8_Archivos/TPN8_EJ01.cpp
@@ -69,10 +69,10 @@ main() {
 }
 
 void CargarMontos(FILE *archivo,int &i) {
-    int n;
+    float n;
 
     printf("\nIngrese el monto total %d (Ingrese 0 para salir): ",i+1);
-    scanf("%d",&n);
+    scanf("%f",&n);
 
     while (n!=0)
     {
